+++ codegen/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `introspection` flag to `Options`.
+- Support generating introspection.
 
 ## [0.6.0] - 2024-06-07